+++ packages/adminui-vue/CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 
 ### Fixed
- - UToolbar: fix helpers.showRecordHistory when useVueTables=false
+ - UToolbar: fix showing of record history in case `adminUI.useVueTables=false` in config
 
 ## [5.22.21] - 2022-05-26
 ## [5.22.20] - 2022-05-22